@@ -1,5 +1,5 @@
 package com.example.tx.entity.email;
 
 public interface EmailSender {
-    void send(String to, String email);
+    void send(String subject, String to, String email);
 }
