@@ -1,75 +1,102 @@
-package com.example.tx.service;
+    package com.example.tx.service;
 
-import com.example.tx.entity.user.AppUser;
-import com.example.tx.entity.registration.token.ConfirmationToken;
-import com.example.tx.repository.AppUserRepository;
-import lombok.AllArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
+    import com.example.tx.entity.user.AppUser;
+    import com.example.tx.entity.registration.token.ConfirmationToken;
+    import com.example.tx.repository.AppUserRepository;
+    import lombok.AllArgsConstructor;
+    import lombok.RequiredArgsConstructor;
+    import lombok.extern.slf4j.Slf4j;
+    import org.springframework.context.annotation.Bean;
+    import org.springframework.security.core.authority.SimpleGrantedAuthority;
+    import org.springframework.security.core.userdetails.UserDetails;
+    import org.springframework.security.core.userdetails.UserDetailsService;
+    import org.springframework.security.core.userdetails.UsernameNotFoundException;
+    import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+    import org.springframework.security.crypto.password.PasswordEncoder;
+    import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
-import java.util.UUID;
+    import java.time.LocalDateTime;
+    import java.util.ArrayList;
+    import java.util.Collection;
+    import java.util.UUID;
 
-@Service
-@AllArgsConstructor
-public class AppUserService implements UserDetailsService {
+    @Service
+    @RequiredArgsConstructor
+    @Slf4j
+    public class AppUserService implements UserDetailsService {
 
-    private final static String USER_NOT_FOUND_MSG =
-            "user with email %s not found";
+        private final static String USER_NOT_FOUND_MSG =
+                "user with email %s not found";
 
-    private final AppUserRepository appUserRepository;
-    private final BCryptPasswordEncoder bCryptPasswordEncoder;
-    private final ConfirmationTokenService confirmationTokenService;
+        private final AppUserRepository appUserRepository;
 
-    @Override
-    public UserDetails loadUserByUsername(String email)
-            throws UsernameNotFoundException {
-        return appUserRepository.findByEmail(email)
-                .orElseThrow(() ->
-                        new UsernameNotFoundException(
-                                String.format(USER_NOT_FOUND_MSG, email)));
-    }
+        private final PasswordEncoder bCryptPasswordEncoder;
 
-    public String signUpUser(AppUser appUser) {
-        boolean userExists = appUserRepository
-                .findByEmail(appUser.getEmail())
-                .isPresent();
+        private final ConfirmationTokenService confirmationTokenService;
 
-        if (userExists) {
-            // TODO check of attributes are the same and
-            // TODO if email not confirmed send confirmation email.
 
-            throw new IllegalStateException("email already taken");
+        @Override
+        public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+            AppUser user = appUserRepository.findByUsername(username);
+
+            if(user == null)
+            {
+                log.error("User not found in database");
+                throw new UsernameNotFoundException(String.format(USER_NOT_FOUND_MSG, username));
+            }else{
+                log.info("User found in the database {}", username);
+            }
+
+            return new org.springframework.security.core.userdetails.User(user.getUsername(), user.getPassword(), user.getAuthorities()); // CHECK IF AUTHORITIES ARE PASSED CORRECTLY
+
+
         }
 
-        String encodedPassword = bCryptPasswordEncoder
-                .encode(appUser.getPassword());
+        public String signUpUser(AppUser appUser) {
+            boolean userExists = appUserRepository
+                    .findByUsername(appUser.getUsername())
+                    != null;
 
-        appUser.setPassword(encodedPassword);
+            if (userExists) {
+                // TODO check of attributes are the same and
+                // TODO if email not confirmed send confirmation email.
 
-        appUserRepository.save(appUser);
+                throw new IllegalStateException("email already taken");
+            }
 
-        String token = UUID.randomUUID().toString();
+           String token = saveUser(appUser);
 
-        ConfirmationToken confirmationToken = new ConfirmationToken(
-                token,
-                LocalDateTime.now(),
-                LocalDateTime.now().plusMinutes(15),
-                appUser
-        );
+                //        TODO: SEND EMAIL
 
-        confirmationTokenService.saveConfirmationToken(
-                confirmationToken);
+            return token;
+        }
 
-//        TODO: SEND EMAIL
+        public String saveUser(AppUser appUser)
+        {
+            String encodedPassword = bCryptPasswordEncoder
+                    .encode(appUser.getPassword());
 
-        return token;
+            appUser.setPassword(encodedPassword);
+
+            appUserRepository.save(appUser);
+
+            String token = UUID.randomUUID().toString();
+
+            ConfirmationToken confirmationToken = new ConfirmationToken(
+                    token,
+                    LocalDateTime.now(),
+                    LocalDateTime.now().plusMinutes(15),
+                    appUser
+            );
+
+            confirmationTokenService.saveConfirmationToken(
+                    confirmationToken);
+
+            return token;
+        }
+
+
+        public int enableAppUser(String email) {
+            return appUserRepository.enableAppUser(email);
+        }
     }
-
-    public int enableAppUser(String email) {
-        return appUserRepository.enableAppUser(email);
-    }
-}
