@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
-
+// TODO remove this redundant class and its references
 @Service
 @AllArgsConstructor
 public class ConfirmationTokenService {
